Rewrite the tools
