+++ libraries/CMakeLists.txt
@@ -24,4 +24,4 @@ add_subdirectory(cosmwasm)
 add_subdirectory(boost)
 add_subdirectory(rt)
 
-file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/meta_refl/include/bluegrass DESTINATION ${BASE_BINARY_DIR}/include/internal)
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/meta_refl/include/bluegrass DESTINATION ${BASE_BINARY_DIR}/include)
