+++ libraries/boost/CMakeLists.txt
@@ -1,1 +1,1 @@
-file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/include/boost DESTINATION ${BASE_BINARY_DIR}/include/internal)
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/include/boost DESTINATION ${BASE_BINARY_DIR}/include)
