+++ libraries/eosiolib/CMakeLists.txt
@@ -1,3 +1,5 @@
+set(CMAKE_CXX_STANDARD 20)
+
 file(GLOB HEADERS "*.hpp"
                   "*.h")
 
@@ -20,9 +22,11 @@ add_library(eosio_cmem
 
 set_target_properties(eosio_malloc PROPERTIES LINKER_LANGUAGE C)
 
-target_include_directories(eosio PUBLIC
-                                 ${CMAKE_SOURCE_DIR}
-                                 ${CMAKE_SOURCE_DIR}/boost/include)
+target_include_directories(eosio
+  PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/contracts/eosio
+    ${CMAKE_CURRENT_SOURCE_DIR}/core/eosio
+    ${CMAKE_CURRENT_SOURCE_DIR}/../boost/include)
 
 target_link_libraries( eosio c++ c )
 
@@ -31,4 +35,5 @@ add_custom_command( TARGET eosio_malloc POST_BUILD COMMAND ${CMAKE_COMMAND} -E c
 add_custom_command( TARGET eosio_dsm POST_BUILD COMMAND ${CMAKE_COMMAND} -E copy $<TARGET_FILE:eosio_dsm> ${BASE_BINARY_DIR}/lib )
 add_custom_command( TARGET eosio_cmem POST_BUILD COMMAND ${CMAKE_COMMAND} -E copy $<TARGET_FILE:eosio_cmem> ${BASE_BINARY_DIR}/lib )
 
-file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../eosiolib DESTINATION ${BASE_BINARY_DIR}/include FILES_MATCHING PATTERN "*.h" PATTERN "*.hpp")
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../eosiolib/contracts/ DESTINATION ${BASE_BINARY_DIR}/include FILES_MATCHING PATTERN "*.h" PATTERN "*.hpp")
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/../eosiolib/core/ DESTINATION ${BASE_BINARY_DIR}/include FILES_MATCHING PATTERN "*.h" PATTERN "*.hpp")
