+++ libraries/eosiolib/contracts/eosio/eosio.hpp
@@ -4,7 +4,7 @@
  */
 #pragma once
 #include "action.hpp"
-#include "../../core/eosio/print.hpp"
+#include "print.hpp"
 #include "map.hpp"
 #include "multi_index.hpp"
 #include "dispatcher.hpp"
