+++ libraries/eosiolib/contracts/eosio/map.hpp
@@ -1,10 +1,10 @@
 #pragma once
-#include "../../core/eosio/context.hpp"
-#include "../../core/eosio/datastream.hpp"
-#include "../../core/eosio/name.hpp"
-#include "../../core/eosio/varint.hpp"
+#include "context.hpp"
+#include "datastream.hpp"
+#include "name.hpp"
+#include "varint.hpp"
 
-#include "../../core/eosio/key_utils.hpp"
+#include "key_utils.hpp"
 
 #include <algorithm>
 #include <cctype>
