+++ libraries/eosiolib/contracts/eosio/multi_index.hpp
@@ -4,10 +4,10 @@
  */
 #pragma once
 
-#include "../../contracts/eosio/action.hpp"
-#include "../../core/eosio/name.hpp"
-#include "../../core/eosio/serialize.hpp"
-#include "../../core/eosio/fixed_bytes.hpp"
+#include "action.hpp"
+#include "name.hpp"
+#include "serialize.hpp"
+#include "fixed_bytes.hpp"
 
 #include <vector>
 #include <tuple>
