+++ modules/BlancExecutableNames.cmake
@@ -6,17 +6,17 @@ set(BLANC_POSTPASS ${BLANC_PREFIX}-pp)
 set(BLANC_AR ${BLANC_PREFIX}-ar)
 set(BLANC_RANLIB ${BLANC_PREFIX}-ranlib)
 
-set(BLANC_C_COMPILER_BACKEND clang)
-set(BLANC_CXX_COMPILER_BACKEND clang++)
-set(BLANC_LINKER_BACKEND wasm-ld)
-set(BLANC_AR_BACKEND llvm-ar)
-set(BLANC_RANLIB_BACKEND llvm-ranlib)
+set(BLANC_C_COMPILER_BACKEND_FALLBACK clang)
+set(BLANC_CXX_COMPILER_BACKEND_FALLBACK clang++)
+set(BLANC_LINKER_BACKEND_FALLBACK wasm-ld)
+set(BLANC_AR_BACKEND_FALLBACK llvm-ar)
+set(BLANC_RANLIB_BACKEND_FALLBACK llvm-ranlib)
 
-set(BLANC_C_COMPILER_BACKEND_PREFERRED ${BLANC_C_COMPILER_BACKEND}-12)
-set(BLANC_CXX_COMPILER_BACKEND_PREFERRED ${BLANC_CXX_COMPILER_BACKEND}-12)
-set(BLANC_LINKER_BACKEND_PREFERRED ${BLANC_LINKER_BACKEND}-12)
-set(BLANC_AR_BACKEND_PREFERRED ${BLANC_AR_BACKEND}-12)
-set(BLANC_RANLIB_BACKEND_PREFERRED ${BLANC_RANLIB_BACKEND}-12)
+set(BLANC_C_COMPILER_BACKEND ${BLANC_C_COMPILER_BACKEND_FALLBACK}-12)
+set(BLANC_CXX_COMPILER_BACKEND ${BLANC_CXX_COMPILER_BACKEND_FALLBACK}-12)
+set(BLANC_LINKER_BACKEND ${BLANC_LINKER_BACKEND_FALLBACK}-12)
+set(BLANC_AR_BACKEND ${BLANC_AR_BACKEND_FALLBACK}-12)
+set(BLANC_RANLIB_BACKEND ${BLANC_RANLIB_BACKEND_FALLBACK}-12)
 
 if(APPLE)
   set(BREW_LLVM_PATH "/usr/local/opt/llvm")
