+++ modules/EosioCDTMacros.cmake.in
@@ -1,9 +1,7 @@
 if (NOT EOSIO_WASM_OLD_BEHAVIOR STREQUAL "On")
    macro(add_contract CONTRACT_NAME TARGET)
       add_executable( ${TARGET} ${ARGN} )
-      target_compile_options( ${TARGET} PUBLIC -abigen )
       get_target_property(BINOUTPUT ${TARGET} BINARY_DIR)
-      target_compile_options( ${TARGET} PUBLIC -abigen_output=${BINOUTPUT}/${TARGET}.abi )
       target_compile_options( ${TARGET} PUBLIC -contract ${CONTRACT_NAME} )
    endmacro()
    macro (target_ricardian_directory TARGET DIR)
@@ -15,9 +13,7 @@ if (NOT EOSIO_WASM_OLD_BEHAVIOR STREQUAL "On")
 else()
    macro(add_contract CONTRACT_NAME TARGET)
       add_executable( ${TARGET}.wasm ${ARGN} )
-      target_compile_options( ${TARGET}.wasm PUBLIC -abigen )
       get_target_property(BINOUTPUT ${TARGET}.wasm BINARY_DIR)
-      target_compile_options( ${TARGET}.wasm PUBLIC -abigen_output=${BINOUTPUT}/${TARGET}.abi )
       target_compile_options( ${TARGET}.wasm PUBLIC -contract ${CONTRACT_NAME} )
    endmacro()
    macro (target_ricardian_directory TARGET DIR)
