+++ modules/blanc-config.cmake
@@ -4,7 +4,8 @@ endif()
 set(EOSIO_CDT_ROOT ${BLANC_ROOT_DIR})
 
 list(APPEND CMAKE_MODULE_PATH ${BLANC_ROOT_DIR}/lib/cmake/@CMAKE_PROJECT_NAME@)
-if(EOSIO_WASM_OLD_BEHAVIOR STREQUAL "On")
+if(NOT EOSIO_WASM_OLD_BEHAVIOR STREQUAL "Off")
+  set(EOSIO_WASM_OLD_BEHAVIOR "On")
   include(EosioWasmToolchain)
 endif()
