+++ plugins/eosio/attrs.cpp
