--- tools/include/blanc/compiler_options.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "./profile.hpp"
