+++ tools/include/eosio/abi.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <iostream>
+#include <set>
 #include <string>
 #include <vector>
 #include <unordered_set>
