+++ tools/include/eosio/abimerge.hpp
@@ -32,11 +32,11 @@ class ABIMerger {
          ret["structs"]  = merge_structs(other);
          ret["actions"]  = merge_actions(other);
          ret["tables"]   = merge_tables(other);
-         ret["kv_tables"] = merge_kv_tables(other);
          ret["ricardian_clauses"]  = merge_clauses(other);
          ret["variants"] = merge_variants(other);
          std::string vers = abi["version"].as<std::string>();
          if (std::stod(vers.substr(vers.size()-3))*10 >= 12) {
+            ret["kv_tables"] = merge_kv_tables(other);
             ret["action_results"] = merge_action_results(other);
          }
          return ret;
