--- tools/src/blanc-cc.cpp.in
@@ -1,95 +1,0 @@
-// Declares clang::SyntaxOnlyAction.
-#include "clang/Frontend/FrontendActions.h"
-#include "clang/Tooling/CommonOptionsParser.h"
-#include "clang/Tooling/Tooling.h"
-#include <iostream>
-#include <sstream>
-
-// Declares llvm::cl::extrahelp.
-#include "llvm/Support/CommandLine.h"
-#include "llvm/Support/FileSystem.h"
-#include "llvm/Support/Program.h"
-#include "llvm/Support/Path.h"
-
-using namespace clang::tooling;
-using namespace llvm;
-
-#define COMPILER_NAME "${BLANC_C_COMPILER}"
-#define LINKER_NAME "${BLANC_LINKER}"
-#define COMPILER_BACKEND "${BLANC_C_COMPILER_BACKEND}"
-#define COMPILER_BACKEND_PREFERRED "${BLANC_C_COMPILER_BACKEND_PREFERRED}"
-#include <compiler_options.hpp>
-
-int main(int argc, const char **argv) {
-   std::vector<std::string> backends = { COMPILER_BACKEND_PREFERRED, COMPILER_BACKEND };
-
-   for (int i=0; i < argc; i++) {
-     if (argv[i] == std::string("-v")) {
-       eosio::cdt::environment::exec_subprogram(backends, {"-v"});
-       return 0;
-     }
-   }
-
-   cl::SetVersionPrinter([](llvm::raw_ostream& os) {
-         os << COMPILER_NAME << " version " << "${VERSION_FULL}" << "\n";
-   });
-   cl::ParseCommandLineOptions(argc, argv, std::string(COMPILER_NAME)+" (Eosio C -> WebAssembly compiler)");
-   Options opts = CreateOptions();
-
-   if (opts.flags.abigen) {
-       llvm::outs() << "Warning, ABI generation is only available with eosio-abigen or eosio-cpp\n";
-   }
-   
-   std::vector<std::string> outputs;
-   for (auto input : opts.inputs) {
-      std::vector<std::string> new_opts = opts.comp_options;
-      SmallString<64> res;
-      llvm::sys::path::system_temp_directory(true, res);
-      std::string tmp_file = std::string(res.c_str())+"/"+input+"-tmp.c";
-      std::string output;
-
-      auto src = SmallString<64>(input);
-      llvm::sys::path::remove_filename(src);
-      std::string source_path = src.str().empty() ? "." : src.str().str();
-      new_opts.insert(new_opts.begin(), "-I" + source_path);
-
-      output = tmp_file+".o";
-
-      new_opts.insert(new_opts.begin(), input);
-
-      if (!opts.flags.link) {
-         output = opts.output_fn.empty() ? "a.out" : opts.output_fn;
-      }
-
-      new_opts.insert(new_opts.begin(), output);
-      new_opts.insert(new_opts.begin(), "-o");
-      outputs.push_back(output);
-
-      if (!eosio::cdt::environment::exec_subprogram(backends, new_opts)) {
-         llvm::sys::fs::remove(tmp_file);
-         return -1;
-      }
-      llvm::sys::fs::remove(tmp_file);
-   } 
-   // then link
-   //
-   if (opts.flags.link) {
-      std::vector<std::string> new_opts = opts.ld_options;
-      for (auto input : outputs) {
-         new_opts.insert(new_opts.begin(), std::string(" ")+input+" ");
-      }
-      if (!eosio::cdt::environment::exec_subprogram(LINKER_NAME, new_opts)) {
-         for (auto input : outputs) {
-            llvm::sys::fs::remove(input);
-         }
-         return -1;
-      }
-      for (auto input : outputs) {
-         llvm::sys::fs::remove(input);
-      }
-      if ( !llvm::sys::fs::exists( opts.output_fn ) ) {
-         return -1;
-      }
-   }
-   return 0;
-}
