--- tools/src/blanc-ranlib.cpp.in
@@ -1,18 +1,0 @@
-#include <eosio/utils.hpp>
-
-#define RANLIB_BACKEND "@BLANC_RANLIB_BACKEND@"
-#define RANLIB_BACKEND_PREFERRED "@BLANC_RANLIB_BACKEND_PREFERRED@"
-
-int main(int argc, char** argv) {
-   std::vector<std::string> backends = { RANLIB_BACKEND_PREFERRED, RANLIB_BACKEND };
-
-   std::vector<std::string> args(argc-1);
-   for (auto i = 1; i < argc; ++i) {
-      args[i-1] = argv[i];
-   }
-
-   if (!eosio::cdt::environment::exec_subprogram(backends, args)) {
-      return -1;
-   }
-   return 0;
-}
